Making scenario generator configurable through element mapping file.
